+++ Scripts/GenerateDocumentation/generateDocumentation.sh
@@ -9,7 +9,6 @@ services=(
   AlchemyLanguageV1
   AlchemyVisionV1
   ConversationV1
-  ConversationV1Experimental
   DialogV1
   DocumentConversionV1
   LanguageTranslatorV2